==> notebooks/02_data_quality_and_preprocessing.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("../data/raw/loan_data_raw.csv")
df.head()


df.info()


df = df.drop_duplicates()


df = df.dropna(subset=[
    'income',
    'loan_amount',
    'Credit_Score',
    'LTV',
    'dtir1'
])


df = df[
    (df['income'] > 0) &
    (df['loan_amount'] > 0) &
    (df['Credit_Score'] >= 300)
]


df.info()


df.to_csv("../data/processed/loan_data_clean.csv", index=False)


df['loan_to_income'] = df['loan_amount'] / df['income']


df['loan_to_income_capped'] = df['loan_to_income'].clip(upper=5)


def credit_score_band(score):
    if score < 580:
        return 'Poor'
    elif score < 670:
        return 'Fair'
    elif score < 740:
        return 'Good'
    else:
        return 'Excellent'

df['credit_score_band'] = df['Credit_Score'].apply(credit_score_band)


df[['loan_to_income_capped', 'credit_score_band']].head()


df.to_csv("../data/processed/loan_data_features.csv", index=False)
